requirements.txt
streamlit==1.39.0
tensorflow==2.19.0
opencv-python==4.10.0.84
pillow==10.4.0
numpy==1.26.4